requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.8.0rc2
Django==4.2.23
django-extensions==3.2.3
djangorestframework==3.15.2
djoser==2.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.10
oauthlib==3.3.1
pillow==10.4.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
